backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
openai==1.51.0
httpx==0.27.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
